core: use button.tsx
